requirements.txt
django-photologue>=2.2

